scf-claude-lab/requirements.txt
pyyaml>=6.0
jsonschema>=4.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0
PyPDF2>=3.0
nltk>=3.8
python-multipart>=0.0.6
